+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -  Prevent content overflow if `max-width: 100%` is used inside Position component
 
+### Fixed
+
+- Accordion component triggers scroll on page load
+
 ## 3.2.4 (December 3, 2019)
 
 ### Added